requirements.txt
deta
fastapi
uvicorn
python-dotenv
python-multipart
passlib
passlib[bcrypt]
python-jose[cryptography]
pycryptodome
pycryptodomex
cryptography